# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BBYO Connect API
  version: 1.0.0
  summary: Secure closed social network backend for BBYO teens and chapter accounts
  description: |
    Production-oriented backend API for BBYO Connect.

    Security and youth safety features:
    - JWT access tokens + rotating refresh tokens
    - Role-based access control
    - Audit logging
    - Reporting and moderation workflows
    - Language filtering hooks

    Realtime channel (Socket.IO):
    - message.created
    - message.edited
    - message.deleted
    - conversation.updated
    - notification.created
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PostIdParam:
      name: postId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional error details
    ApiEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          description: Response payload
        meta:
          description: Pagination or extra metadata
        error:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/ApiError'
...
    Locale:
      type: string
      enum: [en, es, he, fr]
...
    PostCreateRequest:
      type: object
      required: [category, text, language, visibility]
      properties:
        category:
          type: string
          enum: [general, chapter_update, event, leadership, creative, safety]
        text:
          type: string
        language:
          $ref: '#/components/schemas/Locale'
        visibility:
          type: string
          enum: [public, chapter, region, global, private]
        media:
          type: array
          items:
            type: object
            required: [type, url]
            properties:
              type:
                type: string
                enum: [image, video, audio, voice_note, file]
              url:
                type: string
                format: uri
              thumbnailUrl:
                type: string
                format: uri
              duration:
                type: integer
...
    ResolveReportRequest:
      type: object
      required: [status, actionType]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        actionType:
          type: string
          enum: [hide_content, warn_user, suspend_user, ban_user, restore_content, note]
        notes:
          type: string
paths:
...
  /feed:
    get:
      tags: [Feed]
      summary: Feed listing (cursor-based)
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [general, chapter_update, event, leadership, creative, safety]
      responses:
        '200':
          description: Feed page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    post:
      tags: [Feed]
      summary: Create feed post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /feed/{postId}:
    get:
      tags: [Feed]
      summary: Get single post
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    patch:
      tags: [Feed]
      summary: Edit post
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
                category:
                  type: string
                  enum: [general, chapter_update, event, leadership, creative, safety]
                visibility:
                  type: string
                  enum: [public, chapter, region, global, private]
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    delete:
      tags: [Feed]
      summary: Soft delete post
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Post deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /feed/{postId}/reactions:
    post:
      tags: [Feed]
      summary: Add/update post reaction
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reactionType]
              properties:
                reactionType:
                  type: string
                  enum: [like, love, support, clap, celebrate, fire]
      responses:
        '201':
          description: Reaction set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    delete:
      tags: [Feed]
      summary: Remove own reaction
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      responses:
        '200':
          description: Reaction removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /feed/{postId}/comments:
    get:
      tags: [Feed]
      summary: List comments for post
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Comment page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    post:
      tags: [Feed]
      summary: Add comment to post
      parameters:
        - $ref: '#/components/parameters/PostIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, language]
              properties:
                text: { type: string }
                language: { $ref: '#/components/schemas/Locale' }
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
  /reports:
    post:
      tags: [Reports]
      summary: Create content or user report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId, reason]
              properties:
                targetType:
                  type: string
                  enum: [user, post, comment, story, message, event, chapter]
                targetId: { type: string }
                reason: { type: string }
                details: { type: string }
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
    get:
      tags: [Reports]
      summary: List reports for moderators
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [open, under_review, resolved, dismissed]
      responses:
        '200':
          description: Reports page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /reports/{id}/resolve:
    patch:
      tags: [Reports]
      summary: Resolve or dismiss report and create moderation action
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveReportRequest'
      responses:
        '200':
          description: Report resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...